tutorials/tutorial-generated-examples/strands-ai-agent-diagram-and-rules/requirements.txt
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0
strands-agents
strands-agents-tools